Use u16 for tile coordinates
Change tile coordinate type from u8 to u16 across the codebase and update related APIs and docs. Files updated: crate types (tilemap.rs, tmx_parser.rs, image.rs) to use ImageTileCoord = u16 and propagate Tile typing; Python binding (tilemap_wrapper.rs) now exposes the tilemap data as a ctypes.c_uint16 array; Python stub and web/api docs (python/pyxel/__init__.pyi, docs/api-reference.md, web/api-reference/api-reference.json) updated to reflect the pointer format (4 bytes per tile). Changelog entry updated accordingly. Note: this changes the raw data layout (now 4 bytes per tile), so native consumers of Tilemap.data_ptr must be adjusted.

+++ crates/pyxel-core/src/tilemap.rs
@@ -6,7 +6,7 @@ use crate::settings::TILE_SIZE;
 use crate::tmx_parser::parse_tmx;
 use crate::utils::{f32_to_u32, parse_hex_string, simplify_string};
 
-pub type ImageTileCoord = u8;
+pub type ImageTileCoord = u16;
 pub type Tile = (ImageTileCoord, ImageTileCoord);
 
 impl ToIndex for Tile {
